+++ src/Validator/Rules/PossibleFragmentSpreads.php
@@ -13,6 +13,7 @@ use GraphQL\Type\Definition\ObjectType;
 use GraphQL\Type\Definition\Type;
 use GraphQL\Type\Definition\UnionType;
 use GraphQL\Type\Schema;
+use GraphQL\Utils\AST;
 use GraphQL\Utils\TypeInfo;
 use GraphQL\Validator\QueryValidationContext;
 
@@ -146,7 +147,7 @@ class PossibleFragmentSpreads extends ValidationRule
             return null;
         }
 
-        $type = TypeInfo::typeFromAST([$context->getSchema(), 'getType'], $frag->typeCondition);
+        $type = AST::typeFromAST([$context->getSchema(), 'getType'], $frag->typeCondition);
 
         return $type instanceof CompositeType
             ? $type