+++ src/Validator/Rules/QuerySecurityRule.php
@@ -13,6 +13,7 @@ use GraphQL\Type\Definition\FieldDefinition;
 use GraphQL\Type\Definition\HasFieldsType;
 use GraphQL\Type\Definition\Type;
 use GraphQL\Type\Introspection;
+use GraphQL\Utils\AST;
 use GraphQL\Utils\TypeInfo;
 use GraphQL\Validator\QueryValidationContext;
 use InvalidArgumentException;
@@ -137,7 +138,7 @@ abstract class QuerySecurityRule extends ValidationRule
                     $typeCondition = $selection->typeCondition;
                     $fragmentParentType = $typeCondition === null
                         ? $parentType
-                        : TypeInfo::typeFromAST([$context->getSchema(), 'getType'], $typeCondition);
+                        : AST::typeFromAST([$context->getSchema(), 'getType'], $typeCondition);
                     $astAndDefs = $this->collectFieldASTsAndDefs(
                         $context,
                         $fragmentParentType,
@@ -157,7 +158,7 @@ abstract class QuerySecurityRule extends ValidationRule
                         if ($fragment !== null) {
                             $astAndDefs = $this->collectFieldASTsAndDefs(
                                 $context,
-                                TypeInfo::typeFromAST([$context->getSchema(), 'getType'], $fragment->typeCondition),
+                                AST::typeFromAST([$context->getSchema(), 'getType'], $fragment->typeCondition),
                                 $fragment->selectionSet,
                                 $visitedFragmentNames,
                                 $astAndDefs