+++ src/Validator/Rules/VariablesInAllowedPosition.php
@@ -11,6 +11,7 @@ use GraphQL\Language\AST\VariableDefinitionNode;
 use GraphQL\Type\Definition\NonNull;
 use GraphQL\Type\Definition\Type;
 use GraphQL\Type\Schema;
+use GraphQL\Utils\AST;
 use GraphQL\Utils\TypeComparators;
 use GraphQL\Utils\TypeInfo;
 use GraphQL\Utils\Utils;
@@ -52,7 +53,7 @@ class VariablesInAllowedPosition extends ValidationRule
                         // If both are list types, the variable item type can be more strict
                         // than the expected item type (contravariant).
                         $schema = $context->getSchema();
-                        $varType = TypeInfo::typeFromAST([$schema, 'getType'], $varDef->type);
+                        $varType = AST::typeFromAST([$schema, 'getType'], $varDef->type);
 
                         if ($varType !== null && ! $this->allowedVariableUsage($schema, $varType, $varDef->defaultValue, $type, $defaultValue)) {
                             $context->reportError(new Error(