+++ CHANGELOG.md
@@ -44,7 +44,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Align `Utils::suggestionList()` with the reference implementation (#1075)
 - Order schema topologically and according to the user-defined order, affects introspection and printing
 - `GraphQL\Utils\AST::typeFromAST()` now needs a type loader callable instead of the Schema
-- `GraphQL\Utils\TypeInfo::typeFromAST()` now needs a type loader callable instead of the Schema
+- Removed `GraphQL\Utils\TypeInfo::typeFromAST()` (use `GraphQL\Utils\AST::typeFromAST()` instead)
 
 ### Added